A functioning (all tests passing) verstion of the template inheritance feature. Refactored the template library into smaller sub-libraries for easier maintenance.
